requirements.txt
bcrypt==4.2.0
cffi==1.17.1
cryptography==43.0.3
dnspython==2.7.0
numpy==2.1.3
pandas==2.2.3
pycparser==2.22
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
tzdata==2024.2
